+++ Braintree.Tests/CreditCardTest.cs
@@ -329,7 +329,7 @@ namespace Braintree.Tests
             try {
                 gateway.CreditCard.Find(" ");
                 Assert.Fail("Should throw NotFoundException");
-            } catch (NotFoundException ok) {}
+            } catch (NotFoundException) {}
         }
 
         [Test]
@@ -794,5 +794,228 @@ namespace Braintree.Tests
             HashSet<String> uniqueCards = new HashSet<String>(cards);
             Assert.AreEqual(uniqueCards.Count, collection.MaximumCount);
         }
+
+        [Test]
+        public void Prepaid()
+        {
+            Customer customer = gateway.Customer.Create(new CustomerRequest()).Target;
+            CreditCardRequest request = new CreditCardRequest
+            {
+                CustomerId = customer.Id,
+                CardholderName = "John Doe",
+                CVV = "123",
+                Number = Braintree.Tests.CreditCardNumbers.CardTypeIndicators.Prepaid,
+                ExpirationDate = "05/12",
+                Options = new CreditCardOptionsRequest
+                {
+                    VerifyCard = true
+                }
+            };
+
+            CreditCard creditCard = gateway.CreditCard.Create(request).Target;
+            Assert.AreEqual(Braintree.CreditCardPrepaid.YES, creditCard.Prepaid);
+        }
+
+        [Test]
+        public void Commercial()
+        {
+            Customer customer = gateway.Customer.Create(new CustomerRequest()).Target;
+            CreditCardRequest request = new CreditCardRequest
+            {
+                CustomerId = customer.Id,
+                CardholderName = "John Doe",
+                CVV = "123",
+                Number = Braintree.Tests.CreditCardNumbers.CardTypeIndicators.Commercial,
+                ExpirationDate = "05/12",
+                Options = new CreditCardOptionsRequest
+                {
+                    VerifyCard = true
+                }
+            };
+
+            CreditCard creditCard = gateway.CreditCard.Create(request).Target;
+            Assert.AreEqual(Braintree.CreditCardCommercial.YES, creditCard.Commercial);
+        }
+
+        [Test]
+        public void Debit()
+        {
+            Customer customer = gateway.Customer.Create(new CustomerRequest()).Target;
+            CreditCardRequest request = new CreditCardRequest
+            {
+                CustomerId = customer.Id,
+                CardholderName = "John Doe",
+                CVV = "123",
+                Number = Braintree.Tests.CreditCardNumbers.CardTypeIndicators.Debit,
+                ExpirationDate = "05/12",
+                Options = new CreditCardOptionsRequest
+                {
+                    VerifyCard = true
+                }
+            };
+
+            CreditCard creditCard = gateway.CreditCard.Create(request).Target;
+            Assert.AreEqual(Braintree.CreditCardDebit.YES, creditCard.Debit);
+        }
+
+        [Test]
+        public void Healthcare()
+        {
+            Customer customer = gateway.Customer.Create(new CustomerRequest()).Target;
+            CreditCardRequest request = new CreditCardRequest
+            {
+                CustomerId = customer.Id,
+                CardholderName = "John Doe",
+                CVV = "123",
+                Number = Braintree.Tests.CreditCardNumbers.CardTypeIndicators.Healthcare,
+                ExpirationDate = "05/12",
+                Options = new CreditCardOptionsRequest
+                {
+                    VerifyCard = true
+                }
+            };
+
+            CreditCard creditCard = gateway.CreditCard.Create(request).Target;
+            Assert.AreEqual(Braintree.CreditCardHealthcare.YES, creditCard.Healthcare);
+        }
+
+        [Test]
+        public void Payroll()
+        {
+            Customer customer = gateway.Customer.Create(new CustomerRequest()).Target;
+            CreditCardRequest request = new CreditCardRequest
+            {
+                CustomerId = customer.Id,
+                CardholderName = "John Doe",
+                CVV = "123",
+                Number = Braintree.Tests.CreditCardNumbers.CardTypeIndicators.Payroll,
+                ExpirationDate = "05/12",
+                Options = new CreditCardOptionsRequest
+                {
+                    VerifyCard = true
+                }
+            };
+
+            CreditCard creditCard = gateway.CreditCard.Create(request).Target;
+            Assert.AreEqual(Braintree.CreditCardPayroll.YES, creditCard.Payroll);
+        }
+
+        [Test]
+        public void DurbinRegulated()
+        {
+            Customer customer = gateway.Customer.Create(new CustomerRequest()).Target;
+            CreditCardRequest request = new CreditCardRequest
+            {
+                CustomerId = customer.Id,
+                CardholderName = "John Doe",
+                CVV = "123",
+                Number = Braintree.Tests.CreditCardNumbers.CardTypeIndicators.DurbinRegulated,
+                ExpirationDate = "05/12",
+                Options = new CreditCardOptionsRequest
+                {
+                    VerifyCard = true
+                }
+            };
+
+            CreditCard creditCard = gateway.CreditCard.Create(request).Target;
+            Assert.AreEqual(Braintree.CreditCardDurbinRegulated.YES, creditCard.DurbinRegulated);
+        }
+
+        [Test]
+        public void CountryOfIssuance()
+        {
+            Customer customer = gateway.Customer.Create(new CustomerRequest()).Target;
+            CreditCardRequest request = new CreditCardRequest
+            {
+                CustomerId = customer.Id,
+                CardholderName = "John Doe",
+                CVV = "123",
+                Number = Braintree.Tests.CreditCardNumbers.CardTypeIndicators.CountryOfIssuance,
+                ExpirationDate = "05/12",
+                Options = new CreditCardOptionsRequest
+                {
+                    VerifyCard = true
+                }
+            };
+
+            CreditCard creditCard = gateway.CreditCard.Create(request).Target;
+            Assert.AreEqual(Braintree.Tests.CreditCardDefaults.CountryOfIssuance, creditCard.CountryOfIssuance);
+        }
+
+        [Test]
+        public void IssuingBank()
+        {
+            Customer customer = gateway.Customer.Create(new CustomerRequest()).Target;
+            CreditCardRequest request = new CreditCardRequest
+            {
+                CustomerId = customer.Id,
+                CardholderName = "John Doe",
+                CVV = "123",
+                Number = Braintree.Tests.CreditCardNumbers.CardTypeIndicators.IssuingBank,
+                ExpirationDate = "05/12",
+                Options = new CreditCardOptionsRequest
+                {
+                    VerifyCard = true
+                }
+            };
+
+            CreditCard creditCard = gateway.CreditCard.Create(request).Target;
+            Assert.AreEqual(Braintree.Tests.CreditCardDefaults.IssuingBank, creditCard.IssuingBank);
+        }
+
+        [Test]
+        public void NegativeCardTypeIndicators()
+        {
+            Customer customer = gateway.Customer.Create(new CustomerRequest()).Target;
+            CreditCardRequest request = new CreditCardRequest
+            {
+                CustomerId = customer.Id,
+                CardholderName = "John Doe",
+                CVV = "123",
+                Number = Braintree.Tests.CreditCardNumbers.CardTypeIndicators.No,
+                ExpirationDate = "05/12",
+                Options = new CreditCardOptionsRequest
+                {
+                    VerifyCard = true
+                }
+            };
+
+            CreditCard creditCard = gateway.CreditCard.Create(request).Target;
+            Assert.AreEqual(Braintree.CreditCardPrepaid.NO, creditCard.Prepaid);
+            Assert.AreEqual(Braintree.CreditCardCommercial.NO, creditCard.Commercial);
+            Assert.AreEqual(Braintree.CreditCardHealthcare.NO, creditCard.Healthcare);
+            Assert.AreEqual(Braintree.CreditCardDurbinRegulated.NO, creditCard.DurbinRegulated);
+            Assert.AreEqual(Braintree.CreditCardPayroll.NO, creditCard.Payroll);
+            Assert.AreEqual(Braintree.CreditCardDebit.NO, creditCard.Debit);
+        }
+
+        [Test]
+        public void MissingCardTypeIndicators()
+        {
+            Customer customer = gateway.Customer.Create(new CustomerRequest()).Target;
+            CreditCardRequest request = new CreditCardRequest
+            {
+                CustomerId = customer.Id,
+                CardholderName = "John Doe",
+                CVV = "123",
+                Number = Braintree.Tests.CreditCardNumbers.CardTypeIndicators.Unknown,
+                ExpirationDate = "05/12",
+                Options = new CreditCardOptionsRequest
+                {
+                    VerifyCard = true
+                }
+            };
+
+            CreditCard creditCard = gateway.CreditCard.Create(request).Target;
+            Assert.AreEqual(Braintree.CreditCardPrepaid.UNKNOWN, creditCard.Prepaid);
+            Assert.AreEqual(Braintree.CreditCardCommercial.UNKNOWN, creditCard.Commercial);
+            Assert.AreEqual(Braintree.CreditCardHealthcare.UNKNOWN, creditCard.Healthcare);
+            Assert.AreEqual(Braintree.CreditCardDurbinRegulated.UNKNOWN, creditCard.DurbinRegulated);
+            Assert.AreEqual(Braintree.CreditCardPayroll.UNKNOWN, creditCard.Payroll);
+            Assert.AreEqual(Braintree.CreditCardDebit.UNKNOWN, creditCard.Debit);
+            Assert.AreEqual(Braintree.CreditCard.CountryOfIssuanceUnknown, creditCard.CountryOfIssuance);
+            Assert.AreEqual(Braintree.CreditCard.IssuingBankUnknown, creditCard.IssuingBank);
+        }
+
     }
 }