+++ Braintree/Properties/AssemblyInfo.cs
@@ -29,5 +29,5 @@ using System.Runtime.InteropServices;
 //      Build Number
 //      Revision
 //
-[assembly: AssemblyVersion("2.15.1.0")]
-[assembly: AssemblyFileVersion("2.15.1.0")]
+[assembly: AssemblyVersion("2.20.0.2")]
+[assembly: AssemblyFileVersion("2.20.0.2")]