+++ Braintree/TransactionRequest.cs
@@ -31,7 +31,9 @@ namespace Braintree
     {
         public TransactionCreditCardRequest CreditCard { get; set; }
         public Decimal Amount { get; set; }
+        public String Channel { get; set; }
         public String OrderId { get; set; }
+        public Boolean? Recurring { get; set; }
         public String MerchantAccountId { get; set; }
         public String PurchaseOrderNumber { get; set; }
         public CustomerRequest Customer { get; set; }
@@ -84,6 +86,8 @@ namespace Braintree
             if (Amount != 0) builder.AddElement("amount", Amount);
             builder.AddElement("customer-id", CustomerId);
             builder.AddElement("order-id", OrderId);
+            builder.AddElement("channel", Channel);
+            if (Recurring.HasValue) builder.AddElement("recurring", Recurring);
             builder.AddElement("payment-method-token", PaymentMethodToken);
             builder.AddElement("purchase-order-number", PurchaseOrderNumber);
             builder.AddElement("shipping-address-id", ShippingAddressId);